Safer cross-platform path handling
On Windows, we can end up with a mixture of forward and back slashes in
the paths being passed to ffmpeg and atracdenc, which can cause
problems.

Use path.sep instead of hardcoded forward slashes to ensure consistency.

Windows additionally has a longer list of forbidden characters, so ensure
these are replaced with underscores too.